+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-## 1.8 (unreleased)
+
+## 1.8.0 - 2018-09-21
 
 ### Added
 
@@ -14,6 +15,7 @@
 ### Fixed
 
 - Decoder plugin will now remove header when there is no more encoding, instead of setting to an empty array
+
 
 ## 1.7.0 - 2017-11-30